build:update server for pro
